Logo for the Nav bar
Created an Image link to the homepage in the version of a place holder
photo of a polar bear , but it caused some graphical issues with the nav
bar from the center photo that occurs on both pages. It is an issues
probably with my css and i will attempt to fix it.